Abstracting utility functions from main integration tests for remote disks
